Refs #34110 -- Added StorageSettingsMixin.
